+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 ## [in development]
 
 ### Added
+
+- Dev CLI - Add command "go run ./main.go dev getImagesSecurityIssue" to ease getting a small report of HIGH
+  and CRITICAL Security issues reported by quay.io on container images used by the sf-operator.
+
 ### Changed
 
 - Zookeeper version bumped to 3.8.4